Added functionality to print out new records.
- Added functionality to the "Update" button. When it is pressed, the
database is queried for records and forms a two-dimensional array with
the results. Newer records that have not been responded to are displayed
first, followed by older records that have not been responded to.
Records that have a response are showed after. Currently, the contents
of the table are printed out.